New firmware, filaments and Shutdown/Pause working
A new firmware version is now used: 10.4.16.

Added two new filament profiles, PETg and TPU. BEESUPPLY PLA was
slightly modified.

Shutdown and Pause features weren't functional due to the recent changes
in the PrintSplashAutonomous dialog. They should now be working as
before. ShutdownMenu and PauseMenu dialogs are now independent of
PrintSplashAutonomous, that is, when the user requests a pause, the
PrintSplashAutonomous is closed instead of being put on hold, and a new
one is started when the user decides to resume it's print.

Also, added a new COM mode, NO_OK. This allows to receive responses from
commands that were sent from BEESOFT without waiting for the final "ok".
This was implemented due to the fact that the M32 command, in certain
circumstances, wasn't sending this "ok", and this caused BEESOFT to
block for a while waiting for it.