Filaments folder is now a git repository
The filaments folder is now a git local repository that tracks a github
repository dedicated to the filaments. BEESOFT attempts to pull updates
at startup. If the filaments folder isn't a repo, it's created and
changed to the correct branch.

The branch and URL of the repo can be specified in the config.properties
file using the git.filament_repo_url and git.filament_repo_branch keys.